Remove some un-needed module from lazy loading entry point
https://coveord.atlassian.net/browse/JSUI-1954
