+++ package.json
@@ -113,10 +113,11 @@
     "typedoc": "https://github.com/coveord/typedoc/tarball/master",
     "typescript": "^2.5.2",
     "walk": "^2.3.9",
-    "webpack": "^3.6.0",
+    "webpack": "^3.10.0",
     "webpack-dev-server": "^2.9.1",
     "xml2js": "^0.4.17",
-    "yargs": "^4.7.1"
+    "yargs": "^4.7.1",
+    "webpack-bundle-analyzer": "^2.9.1"
   },
   "dependencies": {
     "@material/select": "^0.3.3",
