+++ src/Core.ts
@@ -1,7 +1,3 @@
-declare function require(modname: string);
-// Shim for IE11 promise
-// Will not override native promise in browsers that support them
-require('es6-promise/auto');
 export { underscoreInstance as _ } from './ui/Base/CoveoUnderscore';
 export * from './BaseModules';
 export * from './MiscModules';
