+++ src/TemplatesModules.ts
@@ -1,4 +1,3 @@
-export { TemplateHelpers } from './ui/Templates/TemplateHelpers';
 export { TemplateCache } from './ui/Templates/TemplateCache';
 export { HtmlTemplate } from './ui/Templates/HtmlTemplate';
 export { UnderscoreTemplate } from './ui/Templates/UnderscoreTemplate';
