+++ src/ui/ResultList/ResultList.ts
@@ -43,7 +43,10 @@ import 'styling/_ResultFrame';
 import 'styling/_Result';
 import { InitializationPlaceholder } from '../Base/InitializationPlaceholder';
 import { get } from '../Base/RegisteredNamedMethods';
+import { TemplateComponentOptions } from '../Base/TemplateComponentOptions';
+import { CoreHelpers } from '../Templates/CoreHelpers';
 
+CoreHelpers.exportAllHelpersGlobally(window['Coveo']);
 export interface IResultListOptions {
   resultContainer?: HTMLElement;
   resultTemplate?: Template;
@@ -119,7 +122,7 @@ export class ResultList extends Component {
      * list. This element will then be used as a result container.
      */
     resultContainer: ComponentOptions.buildChildHtmlElementOption(),
-    resultTemplate: ComponentOptions.buildTemplateOption({ defaultFunction: ResultList.getDefaultTemplate }),
+    resultTemplate: TemplateComponentOptions.buildTemplateOption({ defaultFunction: ResultList.getDefaultTemplate }),
 
     /**
      * Specifies the type of animation to display while waiting for a query to return.
