+++ src/utils/DomUtils.ts
@@ -115,8 +115,4 @@ export class DomUtils {
 
     return header;
   }
-
-  static getCurrentScript(): HTMLScriptElement {
-    return <HTMLScriptElement>document.currentScript;
-  }
 }
